feat(nix): Switch to NixOS!

+++ init.bash
@@ -41,10 +41,13 @@ if [ "$(uname -s)" == "Darwin" ]; then
   *) ;;
   esac
 elif ! command -v nixos-version >/dev/null 2>&1; then
-  # if on Linux but not NixOS, install nixGL
-  echo "Installing nixGL"
-  nix-channel --add https://github.com/guibou/nixGL/archive/main.tar.gz nixgl && nix-channel --update
-  nix-env -iA nixgl.auto.nixGLDefault
+  # if NOT on NixOS
+  if ! test -f "/etc/NIXOS"; then
+    # if on Linux but not NixOS, install nixGL
+    echo "Installing nixGL"
+    nix-channel --add https://github.com/guibou/nixGL/archive/main.tar.gz nixgl && nix-channel --update
+    nix-env -iA nixgl.auto.nixGLDefault
+  fi
 fi
 
 echo
